+++ cli/src/dag/dag_execute.rs
@@ -23,7 +23,7 @@ pub(crate) async fn execute_dag(
     command_title!("Executing Nexus DAG '{dag_id}'");
 
     // Load CLI configuration.
-    let conf = CliConf::load().await.unwrap_or_else(|_| CliConf::default());
+    let conf = CliConf::load().await.unwrap_or_default();
 
     // Nexus objects must be present in the configuration.
     let NexusObjects {
@@ -38,13 +38,7 @@ pub(crate) async fn execute_dag(
     // Create wallet context, Sui client and find the active address.
     let mut wallet = create_wallet_context(&conf.sui.wallet_path, conf.sui.net).await?;
     let sui = build_sui_client(&conf.sui).await?;
-
-    let address = match wallet.active_address() {
-        Ok(address) => address,
-        Err(e) => {
-            return Err(NexusCliError::Any(e));
-        }
-    };
+    let address = wallet.active_address().map_err(NexusCliError::Any)?;
 
     // Fetch gas coin object.
     let gas_coin = fetch_gas_coin(&sui, conf.sui.net, address, sui_gas_coin).await?;
@@ -60,7 +54,7 @@ pub(crate) async fn execute_dag(
 
     let mut tx = sui::ProgrammableTransactionBuilder::new();
 
-    match dag::execute(
+    if let Err(e) = dag::execute(
         &mut tx,
         default_sap,
         &dag,
@@ -71,13 +65,10 @@ pub(crate) async fn execute_dag(
         *primitives_pkg_id,
         *network_id,
     ) {
-        Ok(tx) => tx,
-        Err(e) => {
-            tx_handle.error();
+        tx_handle.error();
 
-            return Err(NexusCliError::Any(e));
-        }
-    };
+        return Err(NexusCliError::Any(e));
+    }
 
     tx_handle.success();