+++ cli/src/tool/tool_list.rs
@@ -15,7 +15,7 @@ pub(crate) async fn list_tools() -> AnyResult<(), NexusCliError> {
     command_title!("Listing all available Neuxs tools");
 
     // Load CLI configuration.
-    let conf = CliConf::load().await.unwrap_or_else(|_| CliConf::default());
+    let conf = CliConf::load().await.unwrap_or_default();
 
     // Nexus objects must be present in the configuration.
     let NexusObjects { tool_registry, .. } = get_nexus_objects(&conf)?;