+++ cli/src/tool/tool_unregister.rs
@@ -20,7 +20,7 @@ pub(crate) async fn unregister_tool(
     }
 
     // Load CLI configuration.
-    let conf = CliConf::load().await.unwrap_or_else(|_| CliConf::default());
+    let conf = CliConf::load().await.unwrap_or_default();
 
     // Nexus objects must be present in the configuration.
     let NexusObjects {
@@ -32,13 +32,7 @@ pub(crate) async fn unregister_tool(
     // Create wallet context, Sui client and find the active address.
     let mut wallet = create_wallet_context(&conf.sui.wallet_path, conf.sui.net).await?;
     let sui = build_sui_client(&conf.sui).await?;
-
-    let address = match wallet.active_address() {
-        Ok(address) => address,
-        Err(e) => {
-            return Err(NexusCliError::Any(e));
-        }
-    };
+    let address = wallet.active_address().map_err(NexusCliError::Any)?;
 
     // Fetch gas coin object.
     let gas_coin = fetch_gas_coin(&sui, conf.sui.net, address, sui_gas_coin).await?;
@@ -54,20 +48,17 @@ pub(crate) async fn unregister_tool(
 
     let mut tx = sui::ProgrammableTransactionBuilder::new();
 
-    match tool::unregister(
+    if let Err(e) = tool::unregister(
         &mut tx,
         &tool_fqn,
         &owner_cap,
         tool_registry,
         *workflow_pkg_id,
     ) {
-        Ok(tx) => tx,
-        Err(e) => {
-            tx_handle.error();
+        tx_handle.error();
 
-            return Err(NexusCliError::Any(e));
-        }
-    };
+        return Err(NexusCliError::Any(e));
+    }
 
     tx_handle.success();