+++ sdk/src/test_utils/sui_mocks.rs
@@ -1,4 +1,4 @@
-use crate::sui;
+use crate::{sui, types::NexusObjects};
 
 /// Create a new [`sui::Coin`] with random values.
 pub fn mock_sui_coin(balance: u64) -> sui::Coin {
@@ -26,5 +26,18 @@ pub fn mock_sui_event_id() -> sui::EventID {
     sui::EventID {
         tx_digest: sui::TransactionDigest::random(),
         event_seq: 0,
+    }
+}
+
+/// Create a new [`sui::EventID`] with random values.
+pub fn mock_nexus_objects() -> NexusObjects {
+    NexusObjects {
+        workflow_pkg_id: sui::ObjectID::random(),
+        primitives_pkg_id: sui::ObjectID::random(),
+        interface_pkg_id: sui::ObjectID::random(),
+        network_id: sui::ObjectID::random(),
+        tool_registry: mock_sui_object_ref(),
+        default_sap: mock_sui_object_ref(),
+        gas_service: mock_sui_object_ref(),
     }
 }