+++ sdk/src/types/mod.rs
@@ -1,5 +1,6 @@
 mod json_dag;
 mod nexus_data;
+mod nexus_objects;
 mod serde_parsers;
 mod tool_meta;
 mod type_name;
@@ -7,6 +8,7 @@ mod type_name;
 pub use {
     json_dag::*,
     nexus_data::NexusData,
+    nexus_objects::NexusObjects,
     serde_parsers::*,
     tool_meta::ToolMeta,
     type_name::TypeName,